Fixed issue on the DC migration tool to handle null persistence (like in web migrations)

+++ SanteDB.DisconnectedClient.UI/DcApplicationContext.cs
@@ -344,7 +344,7 @@ namespace SanteDB.DisconnectedClient.UI
                         try
                         {
                             // If the DB File doesn't exist we have to clear the migrations
-                            if (!File.Exists(retVal.ConfigurationManager.GetConnectionString(retVal.Configuration.GetSection<DcDataConfigurationSection>().MainDataSourceConnectionStringName).GetComponent("dbfile")))
+                            if (hasDatabase && !File.Exists(retVal.ConfigurationManager.GetConnectionString(retVal.Configuration.GetSection<DcDataConfigurationSection>().MainDataSourceConnectionStringName).GetComponent("dbfile")))
                             {
                                 retVal.m_tracer.TraceWarning("Can't find the SanteDB database, will re-install all migrations");
                                 retVal.Configuration.GetSection<DcDataConfigurationSection>().MigrationLog.Entry.Clear();
